+++ src/headless/types/plugins/jingle/index.d.ts
@@ -1,4 +1,5 @@
 import Call from './model.js';
 import Calls from './calls.js';
-export { Call, Calls };
+import RTPSession from './rtp.js';
+export { Call, Calls, RTPSession };
 //# sourceMappingURL=index.d.ts.map
