+++ src/headless/types/plugins/jingle/types.d.ts
@@ -14,7 +14,7 @@ export type JingleConversionOptions = {
 export type CallMedia = 'audio' | 'video';
 export type CallDirection = 'incoming' | 'outgoing';
 export type CallState = 'calling' | 'ringing' | 'connecting' | 'active' | 'ended' | 'failed';
-export type CallEndedReason = 'success' | 'declined' | 'cancelled' | 'answered-elsewhere';
+export type CallEndedReason = 'success' | 'declined' | 'cancelled' | 'answered-elsewhere' | 'no-media';
 export type JMIAction = 'propose' | 'ringing' | 'proceed' | 'reject' | 'retract' | 'accept' | 'finish';
 export type CallAttributes = {
     id: string;
